Fix PR review issues: file sizes, circular refs, graceful container recreation
- Extract WorkspaceReferencesMenu into its own file (WorkspaceTree 552→455 LOC)
- Extract slug/naming utilities into workspace-utils.ts (workspace.ts 516→496 LOC)
- Add circular reference detection in addReference (mutual A↔B blocked)
- Filter stale references in getReferences (dangling IDs from deleted workspaces)
- Defer container recreation when active terminals are present instead of
  killing running sessions; notify user that changes apply on next restart
- Add TerminalManager.getWorkspaceTerminalIds() accessor
- Extract persistConfig helper to deduplicate config-write pattern
